done with float

+++ includes/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: mburl <mburl@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/22 11:37:38 by mburl             #+#    #+#             */
-/*   Updated: 2019/11/29 15:47:33 by mburl            ###   ########.fr       */
+/*   Updated: 2019/12/02 13:18:38 by mburl            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -61,7 +61,7 @@ int		handle_flags(char **str, t_flags *flags);
 int		handle_width(char **str, t_flags *flags, va_list args);
 int		handle_precision(char **str, t_flags *flags, va_list args);
 int		handle_length(char **str_c, t_flags *flags);
-int		ft_gcvt(double nb, char **s, int prec);
+int		ft_gcvt(long double nb, char **s, int prec);
 int		print_o(t_flags *flags, va_list args);
 int		print_u(t_flags *flags, va_list args);
 int		print_d(t_flags *flags, va_list args);
@@ -83,6 +83,6 @@ int		ft_pad_nb(t_flags *flags, va_list args, char *base, char *hash_key);
 
 
 
-int			sv_gcvt(double f, char **s, t_flags *flags);
+int			sv_gcvt(long double f, char **s, t_flags *flags);
 
 #endif
